add no web worker support

+++ main.js
@@ -4,7 +4,6 @@
 
     function init() {
         $('#platform').innerHTML = Benchmark.platform;
-        $('#ww').innerHTML = window.Worker ? 'Yes' : 'No';
 
         var btncount = $('#btncount');
         var count = 0;
@@ -12,17 +11,28 @@
             e.preventDefault();
             count++;
             btncount.innerHTML = count;
-        })
+        });
+
+        if (!!window.Worker) {
+            $('#ww').innerHTML = 'Yes';
+            $('#use-worker').checked = true;
+        } else {
+            $('#ww').innerHTML = 'Yes';
+            $('#use-worker').checked = false;
+            $('#use-worker').disabled = true;
+        }
 
         $('#start').addEventListener('click', function(e) {
             e.preventDefault();
-            startTestSuite(8, 12, 4);
+            startTestSuite(8, 20, 4, $('#use-worker').checked);
         });
     }
 
 
-    function startTestSuite(kmin, kmax, step) {
+    function startTestSuite(kmin, kmax, step, useWorker) {
         suite = new Benchmark.Suite('hashcash');
+        $('#stat').textContent = '';
+        $('#result').textContent = '';
 
         for (var i = kmin; i <= kmax; i += step || 1) {
             suite.add({
@@ -31,25 +41,68 @@
                 'defer': true,
                 'minSamples': 32,
                 'onStart': function() {
-                    this._worker = new Worker('worker.js');
-                    this._results = {};
-                    // reduce execution time
-                    if (this.count > 16) {
-                        this.count = 16;
+                    this._useWorker = useWorker;
+                    if (this._useWorker) {
+                        this._worker = new Worker('worker.js');
+                        this._getStamp = this._worker.postMessage.bind(this._worker);
+                    } else {
+                        this._rusha = new Rusha(4 * 1024 * 1024);
+                        this._getStampInner = function(callback, str) {
+                            var k = parseInt(str.split(':')[1], 10) | 0;
+                            var counter = 0;
+                            var test = str + counter.toString(36);
+
+                            while(!this._leading0s(k, this._rusha.digest(test))) {
+                                counter++;
+                                test = str + counter.toString(36);
+                            }
+
+                            setTimeout(function() {
+                                callback({"data": test});
+                            }, 0);
+                        };
+                        this._leading0s = function(k, str) {
+                            var i = 0;
+                            while (k - 4 >= 0) {
+                                if (str.charAt(i) != '0') {
+                                    return false;
+                                }
+                                k -= 4;
+                                i += 1;
+                            }
+                            var last = parseInt(str.charAt(i), 16);
+                            if ((k == 0) ||
+                                (k == 1 && last <= 7) ||
+                                (k == 2 && last <= 3) ||
+                                (k == 3 && last <= 1)) {
+                                return true;
+                            } else {
+                                return false;
+                            }
+                        };
                     }
+
+                    this._results = {};
                 },
                 'setup': function() {
                     var k = this.benchmark.name.slice(2);
                     var deferred = this;
-
-                    this.benchmark._worker.onmessage = (function(e) {
+                    var resolver = function(e) {
                         if (this._original._results[e.data]) {
-                            this._original._results[e.data] += 1;
+                                this._original._results[e.data] += 1;
                         } else {
                             this._original._results[e.data] = 1;
                         }
                         deferred.resolve();
-                    }).bind(this.benchmark);
+                    };
+
+                    if (this.benchmark._useWorker) {
+                        this.benchmark._worker.onmessage = resolver.bind(this.benchmark);
+                    } else {
+                        this.benchmark._getStamp = function(prefix) {
+                            setTimeout(this._getStampInner.bind(this, resolver.bind(this), prefix), 0);
+                        }
+                    }
 
                     this._prefix = [
                         '1',
@@ -60,17 +113,23 @@
                         ''
                     ].join(':');
 
-                    document.querySelector("#stat").innerText = 'Current prefix: ' + this._prefix;
+                    document.querySelector("#stat").textContent = 'Current prefix: ' + this._prefix;
                 },
                 'fn': function() {
-                    this.benchmark._worker.postMessage(this._prefix);
+                    this.benchmark._getStamp(this._prefix);
                 },
                 'teardown': function() {
-                    this.benchmark._worker.onmessage = undefined;
+                    if (this.benchmark._useWorker) {
+                        this.benchmark._worker.onmessage = undefined;
+                    }
                 },
                 'onComplete': function() {
-                    this._worker.terminate();
-                    this._worker = null;
+                    if (this._useWorker) {
+                        this._worker.terminate();
+                        this._worker = null;
+                    } else {
+                        this._rusha = null;
+                    }
                 }
             });
         }
@@ -81,7 +140,7 @@
         });
 
         suite.on('complete', function() {
-            $('#stat').innerText = 'All done! See below:';
+            $('#stat').textContent = 'All done! See below:';
         });
 
         suite.run({'async': true});
@@ -94,13 +153,16 @@
             '(' + 1 / benchmark.hz + ' secs to get one stamp)',
             'stamps:'
         ];
+        var sum = 0;
         for (var stamp in benchmark._results) {
             if (benchmark._results.hasOwnProperty(stamp)) {
                 append.push('   * ' + stamp + ' (' + benchmark._results[stamp] + ' cycles)');
+                sum += parseInt(stamp.split(':')[5], 36);
             }
         }
+        append.push((sum / stamp.length) + ' tries on average to get one stamp')
         append.push("\n");
-        result.innerText += append.join("\n");
+        result.textContent += append.join("\n");
     }
